requirements.txt
openai
python-dotenv
html2text
marker-pdf 